refactored wasccgraph_actor to use interfaces
