--- redisgraph/crates/common/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "wasccgraph-common"
-version = "0.0.2"
-authors = ["Kevin Hoffman <alothien@gmail.com>"]
-edition = "2018"
-homepage = "https://wascc.dev"
-repository = "https://github.com/wascc/redisgraph"
-description = "Common graph database support for waSCC actors and capability providers"
-license = "Apache-2.0"
-documentation = "https://docs.rs/wascc-host"
-readme = "README.md"
-keywords = ["webassembly", "wasm", "redis", "graphdb", "database"]
-categories = ["wasm", "api-bindings"]
-
-[badges]
-maintenance = { status = "actively-developed" }
-
-[dependencies]
-serde = "1.0.111"
-serde_derive = "1.0.111"
